refactoring to separate cleanup and separate code into multiple files
